+++ cloudsmith_cli/core/tests/test_cli_flag_provider.py
@@ -1,6 +1,6 @@
 """Tests for the CLI flag credential provider."""
 
-from cloudsmith_cli.core.credentials import CredentialContext
+from cloudsmith_cli.core.credentials.models import CredentialContext
 from cloudsmith_cli.core.credentials.providers import CLIFlagProvider
